services/db-mcp-server/requirements.txt
fastmcp==2.13.0
databricks-sdk==0.38.0
snowflake-connector-python==3.13.1
google-cloud-bigquery==3.25.0
psycopg2-binary==2.9.9
boto3>=1.34.0
aiohttp==3.12.14
requests==2.32.4
pydantic==2.11.7
python-dotenv>=1.0.0
cryptography>=41.0.0
